packages/tools/skill-3-drive-syncer/requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-dotenv
asyncio
